Editor: Fix tests after file movement and add new conftest file
