+++ spyder/plugins/editor/widgets/editorstack/tests/assets/ipython_file.ipy
